Keepvid Bookmarklet
A bookmarklet that sends a YouTube video to the deep dark processors of Keepvid's website from which it will never return!!!! HAHAHAHAHAHAHAHAHAHA!!!!!! Occasioanl updates will be realeased as needed. Keepvid, you are free to use this if you give me (Ari Meisels) credit. And no offense but your broken bookmarklet sucks.