Fixed bug
Click either of the PRINT buttons (you can click cancel afterwards) and then try to download the spreadsheet. The reason for this bug is because the print window adds a # to the URL after closing. The # is used for jumping to a certain element on the page by using its id. Because it is there, my script adds '&media=excel' after the # so it reads it as 'jump to the element with the id &media=excel'. So because there is none, it does nothing. I am such a terrible programmer :cry: .